Fix models and use Dialog Interface for opening channel.
Serialization of the models, like user or address, change
the type of them object to the generic object type. Added
copy constructors to user, address_book, and address
models to fix this issue. I also added a dialog interface
to make opening a channel much easier. Open channel handler
is still missing verification and the actual gRPC function
to actually open a channel on the LND network.